Added ability to edit character via changes to..
controller (added update action) and view (added edit button and
created update template)

Moved user link to navbar. Added presence validation to character’s
name property. Added two helper methods to determine whether a
character was dead or alive, and a hero or a villain, and display it on
the show page.