Created "happiest" method to find happiest show...
Fixed sort_by_name_length to finally work with
ActiveRecord_CollectionProxy objects and implemented it to sort shows
by length (desc). Modified show controller’s index method in an attempt
to get it to work with an earlier method and decided to keep it, since
it seems like it might be faster.